requirements.txt
fastapi
uvicorn
requests
python-dotenv
PyGithub
google-generativeai
pytest